+++ lib/Swirl.dart
@@ -1,7 +1,7 @@
-library dartling;
+library swirl;
 
 export "src/io/Server.dart" show Server;
 export "src/io/Controller.dart" show Controller;
-export "src/io/Request.dart" show Request;
-export "src/io/Response.dart" show Response;
+export "src/io/entities/Request.dart" show Request;
+export "src/io/entities/Response.dart" show Response;
 export "src/io/Router.dart" show Router, NestedRouter;