+++ test/DartlingServer_test.dart
@@ -3,19 +3,15 @@
 
 library DartlingServer.test;
 
-import 'package:DartlingServer/Dartling.dart';
 import 'package:test/test.dart';
 
 void main() {
   group('A group of tests', () {
-    Awesome awesome;
 
     setUp(() {
-      awesome = new Awesome();
     });
 
     test('First Test', () {
-      expect(awesome.isAwesome, isTrue);
     });
   });
 }